# java/src/main/resources/openapi-chat.yaml
openapi: 3.0.3
info:
  title: Netdisk API - 聊天功能
  description: 网盘系统聊天功能接口文档
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: 本地开发环境
paths:
...
  /api/v2/chat/conversations:
    post:
      tags:
        - 会话管理
      summary: 创建会话
      description: 创建群聊或私聊会话
      operationId: createConversation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConversationRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '400':
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    get:
      tags:
        - 会话管理
      summary: 获取会话列表
      description: 获取当前用户的会话列表，支持分页和搜索
      operationId: listConversations
      security:
        - BearerAuth: []
      parameters:
        - name: keyword
          in: query
          description: 搜索关键字（会话名称）
          schema:
            type: string
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: 每页条数
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
  /api/v2/chat/conversations/{conversationId}:
    get:
      tags:
        - 会话管理
      summary: 获取会话详情
      description: 获取指定会话的详细信息，包括成员列表
      operationId: getConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetailResponse'

    patch:
      tags:
        - 会话管理
      summary: 更新会话
      description: 修改会话名称或头像
      operationId: updateConversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConversationRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationDetailResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          description: 响应数据

    # ==================== 请求模型 ====================
    CreateConversationRequest:
      type: object
      required:
        - conversationType
        - memberUserIds
      properties:
        conversationType:
          type: string
          enum: [group, direct]
          description: 会话类型（group=群聊，direct=私聊）
          example: "group"
        spaceId:
          type: string
          description: 关联空间ID（可选）
          example: "space_xxx"
        name:
          type: string
          description: 群聊名称（最大128字符）
          example: "技术交流群"
        avatarUrl:
          type: string
          description: 群头像URL
          example: "https://xxx.com/avatar.jpg"
        memberUserIds:
          type: array
          description: 成员用户UUID列表（群聊上限500人，超出部分返回failedMembers）
          items:
            type: string
          example: ["user_uuid_1", "user_uuid_2"]
    UpdateConversationRequest:
      type: object
      properties:
        name:
          type: string
          description: 会话名称
          example: "新群名"
        avatarUrl:
          type: string
          description: 头像URL
          example: "https://xxx.com/new_avatar.jpg"
...
    ConversationResponse:
      type: object
      properties:
        id:
          type: string
          description: 会话UUID
        conversationType:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
        status:
          type: string
        ownerUserId:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberInfo'
    ConversationListResponse:
      type: object
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/ConversationSummary'
    ConversationDetailResponse:
      type: object
      properties:
        id:
          type: string
        conversationType:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
        status:
          type: string
        spaceId:
          type: integer
        ownerUserId:
          type: string
        isMuted:
          type: boolean
        muteUntil:
          type: string
        lastMessageAt:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberInfo'
    ConversationSummary:
      type: object
      properties:
        id:
          type: string
        conversationType:
          type: string
        name:
          type: string
        avatarUrl:
          type: string
        status:
          type: string
        lastMessageAt:
          type: string
        isMuted:
          type: boolean
        muteUntil:
          type: string
        unreadCount:
          type: integer
        lastMessage:
          type: object
    MemberInfo:
      type: object
      properties:
        userId:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
        joinedAt:
          type: string
...